refactor: centralize utilities module
